requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn
pillow
streamlit
tqdm
